First test calculating tof with C

+++ include/peakdetect.h
@@ -5,15 +5,15 @@
 #include <stddef.h>
 #include <stdlib.h>
 
-int local_maxima(const uint16_t data[], size_t data_size, uint16_t peaks[], size_t peaks_size);
+int local_maxima(const uint32_t data[], size_t data_size, uint16_t peaks[], size_t peaks_size);
 
 int defrag_peaks(uint16_t peaks[], size_t peaks_size);
 
-int filter_height(const uint16_t data[], size_t data_size, uint16_t peaks[], size_t peaks_size, uint16_t height);
+int filter_height(const uint32_t data[], size_t data_size, uint16_t peaks[], size_t peaks_size, uint16_t height);
 
-int argsort(const uint16_t data[], uint16_t peaks[], size_t peaks_size, uint16_t peaks_order[]);
+int argsort(const uint32_t data[], uint16_t peaks[], size_t peaks_size, uint16_t peaks_order[]);
 
-int filter_distance(const uint16_t data[], size_t data_size, 
+int filter_distance(const uint32_t data[], size_t data_size, 
                     uint16_t peaks[], size_t peaks_size, 
                     uint16_t argsortArray[], size_t distance);
