+++ tests/test_argsort.c
@@ -15,7 +15,7 @@
 #define N_OF_PEAKS 1283
 
 
-uint16_t data[MAX_DATA_POINTS];
+uint32_t data[MAX_DATA_POINTS];
 uint16_t peaks[N_OF_PEAKS];
 uint16_t argsorted_indices[N_OF_PEAKS];
 uint16_t expected_argsorted_indices[N_OF_PEAKS];
